Add helper function to get propertyName
In many cases we need the propertyName and pass it to another function. However in case we use string and rename the property in the future, we almost always forget to update the string.
Introduce helper function that accepts function as argument. The argument function must return the property which name we need.
Add unit tests for the helper method.

Use the new method to simplify setting of deviceLogOptions.