Add settlePromises method
Add settlePromises method that will execute all promises instead of failing when the first promise is rejected. It will be used to replace `Future.wait(futures)`.
The method returns new promise that will be resolved when all promises are resolved. In case any of them is rejected, after all promises are completed, the returned promise will be rejected with an error, describing all errors during the execution.