Fix `unable to resolve callstack error`
When an error is raised we are trying to resolve the callstack, but there's no source map for TypeScript generated methods for `async` (`__awaiter`). This breaks our logic. In such case just print the line in `.js` instead of `.ts` file.
Remove `source-map.js` from `vendor` and use the npm package.