Fix publicApi getters
When `$injector.requirePublic(...)` is used, the $injector creates a getter in the publicApi object for the specified module.
This allows consumers that require the package, to use `require("mobile-cli-lib").devicesService` for example. In this case, the getter should create instance of the DevicesService.
However this has been broken and has been working only when `$injector.requirePublicClass` is used.
Fix the behavior by resolving the instance in the created getter.
Add unit test, so this will not be broken again.