Fix deviceDetectionInterval
It looks like there's some kind of a problem when a Promise is resolved in unref-ed setInterval. Node.js suddenly thinks the process has no more work to do and does not wait for deviceDetectionInterval.
So fix this by calling unref once the Promise is resolved. This allows consecutive executions of setInterval.
This also fixe `<cli name> device` command.