Fix proxy usage on macOS and Linux and fix using npm proxy
Fix using proxy on macOS and Linux - the getCredentials method should not fail - it should return null on OSes that are not supported.
Fix using npm proxy - CLI automatically detects the proxy and uses it for calls to npm registry. However we have not set the rejectUnauthorized variable and all requests to our internal npm proxy were failing. So in such cases set the rejectUnauthorized to the value of strict-ssl npm config variable.