Pass correct args to ios-sim-portable's stopApplication
When Xcode 8 or later is installed, ios-sim-portable will spawn `xcrun simctl terminate <device id> <app id>` in order to stop the application.
For earlier version the command for stopping the app is `killall <app name>`. So pass correct args to the method.
Update ios-sim-portable to latest version, where the change is implemented.