Fix `tns devices` not listing attached iOS devices
When `skipDeviceDetectionInterval` option is not passed to `devicesService`, it does not wait the default timeout (6 seconds) for looking for iOS devices.
It takes some time to detect the devices, but we resolve the promise immediately (as the deviceDetection interval will report the new devices later).
However `tns devices` command is a short-living process and it exits before receiving the iOS devices. Fix the command by passing `skipDeviceDetectionInterval` option.