Return the real transferred files when transferDirectory of device fs is called
When uploading files on devices/emulators, we are executing either `transferDirectory` or `transferFiles` methods of the specific device instance.
However for Android devices, the transferDirectory method has some logic to transfer only modified files (based on shasums from previous execution). So it doesn't really transfer the directory, but just several files.
For LiveSync purposes it is important to know which are the tranferred files, so change the return type of `transferDirectory` method to return the exact files that we've sent to device.

Try-catch the stop of iOS Application as it may fail in some rare cases - in this case just retry it.