Allow setting eqatec-analytics service settings from child classes
Currently the `AnalyticsServiceBase` is starting the eqatec monitor process and the only setting that callers may pass is the API key.
Expose a new protected method that allows starting the monitor with all settings specified. Keep the old private method `start` and when it is called get all required settings from user-settings-service and analytics-settings-service.
Remove code for restarting Analytics monitor when exception should be tracked. We'll no longer use it in {N} CLI, where we had such requirement.