Fix unit tests on Windows
Currently unit tests fail on Windows as the `type` command used in the tests is not a real executable - it is integrated in the `cmd`. So spawning `type` fails and the tests fail as well.
In order to fix this, pass `shell` option to child process, so the child process will spawn CMD and after that type command will succeed.