Recursively look for projectDir when executing hooks
Whenever executing a hook we try and pass the `projectDir` to the hook, however we expect the original method, marked with `@hook` to either have the actual property `projectDir` as an argument or at least have an argument, named `projectData` from which we can get the `projectDir`.
This is incorrect, as the `projectDir` property may be nested in some argument further down.
Recursively traverse all arguments and try and locate the `projectDir` property.