Remove progress indicator when piping http requests
Currently all http request can be piped in httpClient, which is uses when we want to download some file. Each of these pipes is passed through progressStream, so we can print some indication on the same line that something happens.
We do not want this in CI environment. Also it is strange to print message like "Download completed" in the middle of any operation.
So remove the progress indicator (in fact the percentage part of it has been broken for a while). Extract the method in helpers in case we need it in the future.