Introduce connectToPort method to iOSEmulatorServices
Introduce new method - connectToPort to iOSEmulatorServices. This method will try to connect to specified port on users machine (typically this port will be used by NativeScript applications runnin on iOS Simulator) for a specified amount of time. In case it succeeds, the socket of the connection is returned. In case it fails, undefined is returned.