feat(livesync): execute all hooks
Instead of executing the first hook which tries to replace one of CLI's methods, execute them all in succession.
Pass the a reference to the original CLI method to the first hook, then a reference to the first hook's function to the second and so on.
This approach is inspired by objective-C's method swizzling and serves as a similar mechanism. Hooks should decide wether to call the original method or not.