fix(deviceLogs): iOS device logs are not shown when deployOnDevices is used
When the public method `devicesService.deployOnDevices` is used, the device logs for iOS are not shown. The problem is that the method calls `canStartApplication`, which returns false on Windows.
Remove the check and also ensure the set of projectName for the deviceLogProvider will be called no matter of the way we start iOS Applications (through the mounted image or through Safari on device).