Fix `The parameter null is not valid for this command.` error when create command is executed.
Current behaviour:
When `tns create wqeer sadadfad` command is executed, `The parameter null is not valid for this command.` error is thrown.

New behaviour:
When `tns create wqeer sadadfad` command is executed, `The parameter sadadfad is not valid for this command.` error is thrown.

Fixes https://github.com/NativeScript/nativescript-cli/issues/2727