feat: Validate Java compiler version based on runtime version
Update calls to `nativescript-doctor` package to pass projectDir where applicable. This allows the package to verify installed Java Compiler version against the project's Android runtime.
This is required in case the installed Java version is 10 or above and the Androud runtime in the project is older than 4.1.0. In this case the user cannot build this application for Android, but have to update the Android runtime version or downgrade the Java version.