feat: Introduce method for Android port forwarding
Introduce new method for Android port forwarding in the android-process-service. This method is based on already existing ones, but it skips several checks that are not valid for the current case.
Also fix the removing of the port forward when multiple Android devices/emulators are attached - the current command does not work in this case as it does not specify `-s <device>`. Fix this by persisting the information for forwarded ports and devices