fix: Unable to launch application on specific iOS Simulator
The current logic for launching application on specific iOS Simulator reads the device from the `$options` option. This does not work in many cases (for example `tns debug ios` when multiple simulators are running and CLI prompts you to select one.
So add option in the emulator options to specify device and read it from there.