fix: error is shown for Genymotion emulators with old Geny player
In previous versions of Genymotion, trying to execute the `player` command fails with exit code 1, without real error. In order to handle this case, do not fail in case there's nothing on the stderr (or the process errors really bad).
In recent versions of Genymotion, the stderr contains a string that for Logging activities to file even when the command succeeds. So we have ignored this error for the moment.
Improve the message when there's an error.