fix: Android emulator images are not correct
In case the Android Emulator image does not have displayName set in the config.ini file, the displayName that CLI sets fallbacks to `hw.device.name` property.
However, this name is not the one that is visible in other tools, like Android Studio for example. As a user I expect to see the same name as in Android Studio.
In order to fix this, fallback to the AvdId value from config.ini file and if it is missing, fallback to `hw.device.name`.