fix: unable to work with devices with numeric identifiers
In case the device identifier contains only numeric symbols or in case it looks like a number (16089e09 for example), CLI fails to work with it.
The problem is that CLI thinks an index is passed to it, so it removes "one" from the passed number and tries to find the device on the new index.
The other issue is with the `isNumber` helper method which works incorrectly when the string looks like number. Improve the logic in the method and rename it. We cannot parse safely values with exponent (`1e6` for example), but we do not need such functionality.
The new method name and tests reflects the described behavior.
Before using the identifier as an index, check if we'll find device with the specified identifier.